// res/module/edge.schema.ts
// Generated by TSschema from ../edge.json - do not edit, regenerate instead.
/* eslint-disable */
import type { TSSchemaJSONCollection, TSSchemaFragment, TSSchemaType } from "../../src/index.js";

const schema = {
  "$schema": "https://json-schema.org/draft/2020-12/schema",
  "$id": "urn:tsschema:schema:edge",
  "title": "edge",
  "x-vendor": {
    "empty-array": [],
    "empty-object": {},
    "values": [
      true,
      null,
      -1.5,
      "quote\"d"
    ]
  },
  "$defs": {
    "Plain": {
      "type": "object",
      "properties": {
        "kebab-property": {
          "type": "integer"
        },
        "items": {
          "type": "array",
          "prefixItems": [
            {
              "type": "string"
            }
          ]
        }
      },
      "required": [
        "kebab-property"
      ]
    },
    "kebab-name": {
      "enum": [
        "a",
        "b"
      ]
    },
    "string": {
      "type": "string"
    },
    "Schema": {
      "const": "colliding"
    },
    "Models": {
      "type": "boolean"
    }
  }
} as const;
export default schema satisfies TSSchemaJSONCollection;

export type Schema = typeof schema;

export type Models = { [K in keyof Schema['$defs'] & string]: TSSchemaType<TSSchemaFragment<Schema, K>> };
export type Plain = Models["Plain"];